Validity and compatibility of a cry object of class 'rec_unit_cell'

An object of class 'rec_unit_cell' is a named list of length 6. The first three fields are numeric, the last three of class 'angle'.

Usage

check_rec_unit_cell_validity(x, message = FALSE)

Arguments

x

Object of class 'rec_unit_cell'.

message

A logical variable. If TRUE, the function prints a message on the errors, if any (default is FALSE, i.e. no message printed).

Value

ans A logical value. TRUE means that the input is a valid object of class'rec_unit_cell'.

Examples

# Create an object of class 'rec_unit_cell'
x <- create_rec_unit_cell()

# Check its validity
check_rec_unit_cell_validity(x)
#> [1] TRUE

# Now change a field
x$alphar <- 123

# Check validity again
check_rec_unit_cell_validity(x,TRUE)
#> One or more of the alphar, betar, gammar of this object of class 'rec_unit_cell' are not of class 'angle'.
#> [1] FALSE