Default method for generic "create_merged_reflections"

This method is an alternative call to 'merged_reflections'. .

Usage

# Default S3 method
create_merged_reflections(
  ruc = NULL,
  csym = NULL,
  records = NULL,
  dtypes = NULL,
  ...
)

Arguments

ruc

An object of class 'rec_unit_cell'.

csym

An object of class 'cryst_symm'.

records

A data frame containing all reflections coming from the x-ray data collection on the crystal. This data frame must include at least the three Miller indices, H, K, L (of dtype "H").

dtypes

A character vector whose length is the same as the number of columns in 'records'. One character (a capital letter) is associated with each type of data. For example, a Miller index is of dtype "H"; a structure amplitude is of dtype "F"; an anomalous difference is of dtype "D"; etc (see details later).

...

Additional arguments passed to the create_merged_reflections methods.

Value

An object of class "merged_reflections". It is a named list of length 4 whose names are:

ruc

An object of class "rec_unit_cell".

csym

An object of class "cryst_symm".

records

A data frame containing the data.

dtypes

A character vector containing the type of data (Miller indices, structure factors, etc).

See also

Examples

# Create merged data for a cubic (10 angstrom) unit cell
# of space group P 2 3. Data up to 5 angstroms resolution
mrefs <- create_merged_reflections()
print(mrefs)
#> This is an object of class 'merged_reflections'
#> 
#> Its unit cell has sides:
#>       10.000 ,     10.000 ,     10.000    angstroms.
#> And angles:
#>       90.000 ,     90.000 ,     90.000    degrees.
#> 
#> Its symmetry is represented by space group: P 2 3.
#> These are the first few data lines:
#> 
#>     H  K  L         S
#> 1   0  0 -2 0.2000000
#> 2  -1 -1 -1 0.1732051
#> 3   0 -1 -1 0.1414214
#> 4   1 -1 -1 0.1732051
#> 5  -1  0 -1 0.1414214
#> 6   0  0 -1 0.1000000
#> 7   1  0 -1 0.1414214
#> 8  -1  1 -1 0.1732051
#> 9   0  1 -1 0.1414214
#> 10  1  1 -1 0.1732051