Reciprocal unit cell from a 'cryst_symm' object

Method to create an object of class "rec_unit_cell" starting from an object of class 'cryst_symm'.

Usage

# S3 method for class 'cryst_symm'
create_rec_unit_cell(ar, ...)

Arguments

ar

An object of class 'cryst_symm'.

...

Additional arguments passed to the create_rec_unit_cell.

Value

An object of class "rec_unit_cell". It is a named list of length 6 whose last three slots are of class 'angle'. The cell parameters are calculated from those of the corresponding unit cell. The default unit cell parameters are a=10, b=20, c=15, alpha=70, beta=80, gamma=100. When constrains due to symmetry are required, b and c might be equaled to a, alpha, beta and gamma might be set to 90, gamma might be set to 120 and the three angles might be set equal to each other.

Details

The symmetry of a space group imposes constrains on the parameters of unit cells. For example, the cubic group P 2 3 means that all cell sides have to be equal and all angles have to be equal to 90 degrees. This function suggests the appropriate reciprocal cell compatible with the given space group.

Examples

# Symmetry "C 1 2/c 1"
csym <- cryst_symm("C 1 2/c 1")

# Reciprocal unit_cell
ruc <- create_rec_unit_cell(csym)
print(ruc)
#> This is an object of class 'rec_unit_cell'
#> Its sides are:
#>        0.102 ,      0.050 ,      0.068    1/angstroms.
#> Its angles are:
#>       90.000 ,    100.000 ,     90.000    degrees.