Crystal system corresponding to given space group.

Crystal system corresponding to given space group.

Usage

crystal_system(gn)

Arguments

gn

A natural integer (1,2,3,...). the space group number.

Value

A character string, the name of the crystal system associated with the given space group. If the input integer does not correspond any existing space group, the function returns NULL and throws a warning.

Examples

# P1 is part of the TRICLINIC system
crystal_system(1)
#> [1] "TRICLINIC"

# The object returned is a string
csys <- crystal_system(1)
class(csys)
#> [1] "character"