Deplete systematic absences

Remove systematically-absent reflections from a data frame in which Miller indices are in the first three columns. The systematically-absent reflections are determined by the specific space group.

Usage

deplete_systematic_absences(hkl, SG)

Arguments

hkl

A data frame with first three columns H, K, L corresponding to the three Miller indices. This is normally the 'record' data frame in an object of class "merged_reflections".

SG

A character. The extended Hermann-Mauguin symbol of the crystallographic space group.

Value

hkl The same data frame acquired from input, depleted of all systematic absences.

Details

Crystallography symmetry forces constraints on data from x-ray diffraction. One of these constraints consists in the full cancellation of reflections with certain Miller indices. It is said that the reflection with that specific Miller index is systematically absent. For example, in data corresponding to a crystal with space group C 2, general reflections like (h,k,l) must obey h+k=2n (even number). Thus, the Miller indices (2,3,1) are a systematic absence because 2+3=5 (odd).

Examples

# C 2 monoclinic space group
SG <-"C 1 2 1"

# Create an arbitrary cell compatible with C 2
uc <- unit_cell(10,15,10,90,110,90)

# Crete the related reciprocal cell
ruc <- create_rec_unit_cell(uc)

# Create a full data frame of Miller indices
hkl <- expand.grid(H=-4:4,K=-4:4,L=-4:4)

# Get rid of systematic absences
new_hkl <- deplete_systematic_absences(hkl,SG)

# Compare first 10 items of original and depleted arrays
hkl[1:10,]
#>     H  K  L
#> 1  -4 -4 -4
#> 2  -3 -4 -4
#> 3  -2 -4 -4
#> 4  -1 -4 -4
#> 5   0 -4 -4
#> 6   1 -4 -4
#> 7   2 -4 -4
#> 8   3 -4 -4
#> 9   4 -4 -4
#> 10 -4 -3 -4
new_hkl[1:10,]
#>     H  K  L
#> 1  -4 -4 -4
#> 3  -2 -4 -4
#> 5   0 -4 -4
#> 7   2 -4 -4
#> 9   4 -4 -4
#> 11 -3 -3 -4
#> 13 -1 -3 -4
#> 15  1 -3 -4
#> 17  3 -3 -4
#> 19 -4 -2 -4