Reads and output a CIF file

Reads and output a CIF file

Usage

readCIF(filename, message = FALSE)

Arguments

filename

A character string. The path to a valid CIF file.

message

A logical variable. If TRUE (default) the function prints a message highlighting what is included in the cif file.

Value

A named list. Each name correspond to a valid field in the cif.

Examples

datadir <- system.file("extdata",package="cry")
filename <- file.path(datadir,"AMS_DATA.cif")
lCIF <- readCIF(filename)
print(names(lCIF))
#> [1] "HEADER" "SYMM"   "REFL"   "COOR"   "ANISO"  "SYMB"   "GEOM"  
print(lCIF$INTRO$CELL)
#> NULL
print(lCIF$INTRO$HALL)
#> NULL
print(lCIF$INTRO$HM)
#> NULL
print(lCIF$SYMM)
#> [[1]]
#>  [1] "_space_group_symop_operation_xyz" "'x,y,z'"                         
#>  [3] "'2/3+x,1/3+y,1/3+z'"              "'1/3+x,2/3+y,2/3+z'"             
#>  [5] "'-x,-x+y,-z'"                     "'2/3-x,1/3-x+y,1/3-z'"           
#>  [7] "'1/3-x,2/3-x+y,2/3-z'"            "'y,x,-z'"                        
#>  [9] "'2/3+y,1/3+x,1/3-z'"              "'1/3+y,2/3+x,2/3-z'"             
#> [11] "'x-y,-y,-z'"                      "'2/3+x-y,1/3-y,1/3-z'"           
#> [13] "'1/3+x-y,2/3-y,2/3-z'"            "'-y,x-y,z'"                      
#> [15] "'2/3-y,1/3+x-y,1/3+z'"            "'1/3-y,2/3+x-y,2/3+z'"           
#> [17] "'-x+y,-x,z'"                      "'2/3-x+y,1/3-x,1/3+z'"           
#> [19] "'1/3-x+y,2/3-x,2/3+z'"           
#>