Reads and output an mmCIF file

Reads and output an mmCIF file

Usage

readmm_CIF(filename, message = FALSE)

Arguments

filename

A character string. The path to a valid CIF file.

message

A logical variable. If TRUE (default) the function prints a message highlighting what is included in the cif file.

Value

A named list. Each name correspond to a valid field in the cif.

Examples

datadir <- system.file("extdata",package="cry")
filename <- file.path(datadir,"3syu.cif")
lCIF <- readmm_CIF(filename)
print(names(lCIF))
#> [1] "HEADER"       "REFL"         "EXP_DETAILS"  "STRU_DETAILS" "SOFTWARE"    
#> [6] "VAL_DETAILS" 
print(lCIF$HEADER$Entry)
#> [1] "3SYU"
print(lCIF$HEADER$Symmtery)
#> $VAL
#> NULL
#> 
#> $STD
#> NULL
#> 
print(lCIF$HEADER$CELL)
#>                 KEY   VAL
#> 1          entry_id  3SYU
#> 2          length_a 88.69
#> 3          length_b 88.69
#> 4          length_c 88.69
#> 5       angle_alpha    90
#> 6        angle_beta    90
#> 7       angle_gamma    90
#> 8             Z_PDB    24
#> 9  pdbx_unique_axis     ?
#> 10     length_a_esd     ?
#> 11     length_b_esd     ?
#> 12     length_c_esd     ?
#> 13  angle_alpha_esd     ?
#> 14   angle_beta_esd     ?
#> 15  angle_gamma_esd     ?
print(lCIF$EXP_DETAILS$CRYSTAL_CON$VAL)
#> NULL