Coefficients for scattering factors

The nine coefficients of the four-gaussians approximation for the scattering factors

Usage

sfcoeffs(atom_name)

Arguments

atom_name

A character string. The symbol of the atom whose scattering factor's coefficients are needed (see list of available names in details).

Value

A named list whose three components are vectors and scalars containing the coefficients:

a

Vector of length 4 containing the \(a_j\)

b

Vector of length 4 containing the \(b_j\)

c

Single number, the value of \(c\)

Details

A scattering factor is a function of \(\sigma\equiv\sin(\theta)/\lambda\), which is equal to \(s/2\), with \(s=1/d\) and \(d\) the resolution of the given reflection. The scattering form is satisfactorily modelled as sum of four gaussians: $$ a_1\exp(-b_1 \sigma^2) + a_2\exp(-b_2 \sigma^2) + a_3\exp(-b_3 \sigma^2) + a_4\exp(-b_4 \sigma^2) + c $$ This function returns the coefficients as a named list with three components, \(a=(a_1,a_2,a_3,a_4)\), \(b=(b_1,b_2,b_3,b_4)\) and \(c=c\). The possible atom names (also included are some ionic species) are here summarized.

H

H-1

He

Li

Li+1

Be

Be+2

B

C

Cv

N

O

O-1

F

F-1

Ne

Na

Na+1

Mg

Mg+2

Al

Al+3

Si

Siv

Si+4

P

S

Cl

Cl-1

Ar

K

K+1

Ca

Ca+2

Sc

Sc+3

Ti

Ti+2

Ti+3

Ti+4

V

V+2

V+3

V+5

Cr

Cr+2

Cr+3

Mn

Mn+2

Mn+3

Mn+4

Fe

Fe+2

Fe+3

Co

Co+2

Co+3

Ni

Ni+2

Ni+3

Cu

Cu+1

Cu+2

Zn

Zn+2

Ga

Ga+3

Ge

Ge+4

As

Se

Br

Br-1

Kr

Rb

Rb+1

Sr

Sr+2

Y

Y+3

Zr

Zr+4

Nb

Nb+3

Nb+5

Mo

Mo+3

Mo+5

Mo+6

Tc

Ru

Ru+3

Ru+4

Rh

Rh+3

Rh+4

Pd

Pd+2

Pd+4

Ag

Ag+1

Ag+2

Cd

Cd+2

In

In+3

Sn

Sn+2

Sn+4

Sb

Sb+3

Sb+5

Te

I

I-1

Xe

Cs

Cs+1

Ba

Ba+2

La

La+3

Ce

Ce+3

Ce+4

Pr

Pr+3

Pr+4

Nd

Nd+3

Pm

Pm+3

Sm

Sm+3

Eu

Eu+2

Eu+3

Gd

Gd+3

Tb

Tb+3

Dy

Dy+3

Ho

Ho+3

Er

Er+3

Tm

Tm+3

Yb

Yb+2

yb+3

Lu

Lu+3

Hf

Hf+4

Ta

Ta+5

W

W+6

Re

Os

Os+4

Ir

Ir+3

Ir+4

Pt

Pt+2

Pt+4

Au

Au+1

Au+3

Hg

Hg+1

hg+2

Tl

Tl+1

Tl+3

Pb

Pb+2

Pb+4

Bi

Bi+3

Bi+5

Po

At

Rn

Fr

Ra

Ra+2

Ac

Ac+3

Th

Th+4

Pa

U

U+3

U+4

U+6

Np

Np+3

Np+4

Np+6

Pu

Pu+3

Pu+4

Pu+6

Am

Cm

Bk

Cf

Examples

# Coefficients for the carbon atom
aname <- "C"
lcoeffs <- sfcoeffs(aname)
#> Error in sfcoeffs(aname): could not find function "sfcoeffs"
print(lcoeffs)
#> Error in print(lcoeffs): object 'lcoeffs' not found