Cell parameter constrains from symmetry

This function returns a set of constrains, as string character expressions, imposed by the specific symmetry group on the given unit cell.

Usage

symm_to_cell_const(SG)

Arguments

SG

A character string indicating the extended Hermann-Mauguin symbol for the space group.

Value

vcons A character vector. Each component is a string, like 'alpha=90' or 'a=b', that describes the type of constrain to be applied to a unit cell of a crystal structure with given space group symmetry (see above).

Details

Space group symmetry imposes certain constraints on the values that unit cell parameters can take. For example, the symmetry represented by the monoclinic space group of extended Hermann-Mauguin symbol "P 1 2 1" is compatible with a unit cell in which alpha=gamma=90.

There is just a handful of constrains for unit cells. Here they are indicated with the following set of specific strings:

'No constrains' Like in a triclinic cell.

'alpha=90' The alpha angle is fixed at 90 degrees.

'beta=90' The beta angle is fixed at 90 degrees.

'gamma=90' The gamma angle is fixed at 90 degrees.

'gamma=120' The gamma angle is fixed at 120 degrees.

'alpha=beta=gamma' The three angle have the same value, different from 90 degrees.

'a=b' Cell side a is equal to cell side b.

'a=b=c' The three cell sides are equal.

Examples

# P 1 1 2 (group number 3) corresponds to setting 2
SG <- translate_SG(3,set=2)

# Constrains for this symmetry
stmp <- symm_to_cell_const(SG)
#> Input needs to be an extended Hermann-Mauguin symbol.
print(stmp)
#> NULL

# R 3 (rombohedral setting)
stmp <- symm_to_cell_const("R 3 :R")
print(stmp)
#> [1] "a=b=c"            "alpha=beta=gamma"